Refactor application structure
* Delete password input, since it was unused
* Remove unnused code
* Split cómmunication and game logic
* move template loading into one place
